Implémentation du scénario 3 (cartographie du réseau) lien:http://localhost:9090/admin/trajets

+++ src/main/java/org/example/controller/CheminDeFerController.java
@@ -27,7 +27,7 @@ public class CheminDeFerController {
     }
 
     @GetMapping
-    public List<CheminDeFerDAO> listerParDepart(@RequestParam Long departId) {
+    public List<CheminDeFerDAO> listerParDepart(@RequestParam Integer departId) {
         return service.listerParDepart(departId);
     }
 }
